requirements.txt
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
xgboost==1.7.6
boto3==1.28.57
sagemaker==2.177.0
awscli==1.29.57
mlflow==2.7.1
great-expectations==0.17.12
pandera==0.16.1
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
prometheus-client==0.17.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
moto==4.2.5
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
black==23.9.1
flake8==6.1.0
mypy==1.5.1
bandit==1.7.5
